# backend-v2/src/app.ts
/*
 * ⚠️  WARNING: GENERATED CODE - DO NOT MODIFY ⚠️
 * 
 * This file is automatically generated by the backend code generator.
 * Any manual changes will be overwritten when the generator runs.
 * 
 * To make changes, modify the generator configuration files in:
 * /tools/backend-tools/backend-generator-v2/
 * 
 * Then re-run the generator to update this file.
 */

import express from 'express';
import cors from 'cors';
import dotenv from 'dotenv';
import { setupSwagger } from './config/swagger';

// Load environment variables
dotenv.config();

const app = express();
const PORT = process.env.PORT || 3001;

// Request logging middleware
app.use((req, res, next) => {
  console.log(`📥 ${new Date().toISOString()} - ${req.method} ${req.path}`);
  if (req.body && Object.keys(req.body).length > 0) {
    console.log(`   Body:`, JSON.stringify(req.body, null, 2));
  }
  if (req.query && Object.keys(req.query).length > 0) {
    console.log(`   Query:`, JSON.stringify(req.query, null, 2));
  }
  next();
});

// Middleware
app.use(cors());
app.use(express.json());
app.use(express.urlencoded({ extended: true }));

// Root route redirect to Swagger documentation
app.get('/', (req, res) => {
  res.redirect('/api-docs');
});

// Health check endpoint
app.get('/health', (req, res) => {
  res.json({ status: 'OK', timestamp: new Date().toISOString() });
});

// Setup Swagger documentation
setupSwagger(app);

// Import route handlers
import AuthRouter from './routes/auth';
import AchievementRouter from './routes/achievements';
import ContentRouter from './routes/content';
import GameRouter from './routes/games';
import HistoricGameRouter from './routes/historicGames';
import OpeningRouter from './routes/openings';
import PuzzleAttemptRouter from './routes/puzzleAttempts';
import PuzzleSourceRouter from './routes/puzzleSources';
import PuzzleRouter from './routes/puzzles';
import UserAchievementRouter from './routes/userAchievements';
import UserContentProgreRouter from './routes/userContentProgress';
import UserProfileRouter from './routes/userProfiles';
import UserSessionRouter from './routes/userSessions';
import UserRouter from './routes/users';

// Register API routes
app.use('/api/auth', AuthRouter);
app.use('/api/achievements', AchievementRouter);
app.use('/api/content', ContentRouter);
app.use('/api/games', GameRouter);
app.use('/api/historic-games', HistoricGameRouter);
app.use('/api/openings', OpeningRouter);
app.use('/api/puzzle-attempts', PuzzleAttemptRouter);
app.use('/api/puzzle-sources', PuzzleSourceRouter);
app.use('/api/puzzles', PuzzleRouter);
app.use('/api/user-achievements', UserAchievementRouter);
app.use('/api/user-content-progress', UserContentProgreRouter);
app.use('/api/user-profiles', UserProfileRouter);
app.use('/api/user-sessions', UserSessionRouter);
app.use('/api/users', UserRouter);

// Response logging middleware
app.use((req, res, next) => {
  const originalSend = res.send;
  res.send = function(data) {
    console.log(`📤 ${new Date().toISOString()} - ${req.method} ${req.path} - ${res.statusCode}`);
    if (res.statusCode >= 400) {
      console.log(`   Error Response:`, data);
    }
    return originalSend.call(this, data);
  };
  next();
});

// Error handling middleware
app.use((err: any, req: express.Request, res: express.Response, next: express.NextFunction) => {
  console.error('❌ Internal Error:', err);
  console.error('   Stack:', err.stack);
  res.status(500).json({ error: 'Internal server error' });
});

// 404 handler
app.use('*', (req, res) => {
  res.status(404).json({ error: 'Route not found' });
});

if (require.main === module) {
  app.listen(PORT, () => {
    console.log(`🚀 Server running on port ${PORT}`);
    console.log(`📖 API Documentation: http://localhost:${PORT}/api-docs`);
    console.log(`🏠 Root URL redirects to documentation`);
  });
}

export default app;